Implement non-interactive mode for setup and count
In addition to the existing interactive mode this patch adds a
non-interactive mode that allows the user to setup `gh` with the scopes
required to run `gh-notifications`. It allso adds a `--count` and `-c`
option that instructs the program to just print out the number of
notifications which can then be used by other programs and scripts.